pyproject.toml
[project]
name = "goodenough"
version = "0.1.0"
description = "AI-assisted decision-making tool for Arknights"
readme = "README.md"
requires-python = ">=3.10, <=3.11"
dependencies = [
    "adbutils>=1.2.15",
    "jupyter>=1.1.1",
    "numpy>=2.2.5",
    "opencv-python",
    "pillow>=11.2.1",
    "pytesseract>=0.3.13",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "scrcpy-client>=0.4.7",
    "textual-image>=0.8.2",
    "ultralytics>=8.3.120",
]
